chore: color variable 추가, colors와 medias를 개별 파일로 분리
