+++ src/styles/colors.ts
@@ -1,0 +1,22 @@
+export const colors = {
+  bgWhite: "#fff",
+  bgGray: "#f0f0f0",
+  bgGray1: "#f8f8f8",
+  bgGrayDark: "#484848",
+  bgBlack: "#1b1b1b",
+  bgHeader: "#ffffff80",
+
+  ftWhite: "#f8f8f8",
+  ftBlack: "#1b1b1b",
+  ftGray: "#d9d9d9",
+  ftGrayDark: "#8b8b8b",
+  ftGrayDark2: "#434343",
+  ftGrayDark3: "#2A2A2A",
+  ftBlue: "#0275ff",
+
+  bdGray: "#e0e0e0",
+  bdGrayDark: "#d9d9d9",
+
+  boxShadow: "rgba(0,0,0,0.4)",
+  boxShadowLight: "rgba(0,0,0,0.2)",
+};
